// src/core/api/generated/spring/schemas/SystemConfigAttributeDTOAttributeType.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type SystemConfigAttributeDTOAttributeType = typeof SystemConfigAttributeDTOAttributeType[keyof typeof SystemConfigAttributeDTOAttributeType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SystemConfigAttributeDTOAttributeType = {
  ENUM: 'ENUM',
  STRING: 'STRING',
  NUMBER: 'NUMBER',
  BOOLEAN: 'BOOLEAN',
} as const;